feat: Implement domain use cases and refactor user data models and repository interfaces into the domain layer.
