+++ composeApp/src/commonMain/kotlin/com/initium/assignment/data/local/di/DataLocalModule.kt
@@ -4,8 +4,8 @@ import com.initium.assignment.data.local.contracts.IUserLocal
 import com.initium.assignment.data.local.database.createDatabase
 import com.initium.assignment.data.local.database.sqlDriverFactory
 import com.initium.assignment.data.local.implementations.UserLocalImpl
-import com.initium.assignment.data.repository.local.contracts.IUserLocalRepo
-import com.initium.assignment.data.repository.local.implementaions.UserLocalRepoImpl
+import com.initium.assignment.domain.repository.IUserLocalRepo
+import com.initium.assignment.data.repository.local.implementations.UserLocalRepoImpl
 import org.koin.dsl.module
 
 val dataLocalModule = module {
