+++ composeApp/src/commonMain/kotlin/com/initium/assignment/domain/usecase/ClearUserCacheUseCase.kt
@@ -1,0 +1,14 @@
+package com.initium.assignment.domain.usecase
+
+import com.initium.assignment.domain.repository.IUserLocalRepo
+
+class ClearUserCacheUseCase(
+    private val userLocalRepo: IUserLocalRepo
+) {
+    /**
+     * Remove all users from local db.
+     */
+    operator fun invoke() {
+        userLocalRepo.removeAllUsers()
+    }
+}
