+++ composeApp/src/commonMain/kotlin/com/initium/assignment/ui/app/user/mvi/UserState.kt
@@ -1,7 +1,7 @@
 package com.initium.assignment.ui.app.user.mvi
 
-import com.initium.assignment.model.ListDataStruct
-import com.initium.assignment.model.User
+import com.initium.assignment.domain.model.ListDataStruct
+import com.initium.assignment.domain.model.User
 import com.initium.assignment.ui.core.mvi.UiState
 
 data class UserState(
