+++ composeApp/src/commonMain/kotlin/com/initium/assignment/ui/app/userdetail/mvi/UserDetailState.kt
@@ -1,6 +1,6 @@
 package com.initium.assignment.ui.app.userdetail.mvi
 
-import com.initium.assignment.model.User
+import com.initium.assignment.domain.model.User
 import com.initium.assignment.ui.core.mvi.UiState
 
 data class UserDetailState(
