+++ gradle.properties
@@ -1,10 +1,16 @@
+#Gradle
+org.gradle.jvmargs=-Xmx4G
+org.gradle.caching=true
+org.gradle.configuration-cache=true
+org.gradle.daemon=true
+org.gradle.parallel=true
+
 #Kotlin
 kotlin.code.style=official
-kotlin.daemon.jvmargs=-Xmx2048M
-
-#Gradle
-org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8
+kotlin.daemon.jvmargs=-Xmx4G
+kotlin.native.binary.gc=cms
+kotlin.incremental.wasm=true
 
 #Android
-android.nonTransitiveRClass=true
 android.useAndroidX=true
+android.nonTransitiveRClass=true
