+++ gradle/libs.versions.toml
@@ -1,9 +1,9 @@
 [versions]
-agp = "8.13.2"
+agp = "8.11.2"
 android-compileSdk = "36"
 android-minSdk = "24"
 android-targetSdk = "36"
-androidx-activityCompose = "1.12.2"
+androidx-activityCompose = "1.12.4"
 androidx-appcompat = "1.7.1"
 androidx-constraintlayout = "2.2.1"
 androidx-core-ktx = "1.17.0"
@@ -11,22 +11,22 @@ androidx-espresso-core = "3.7.0"
 androidx-lifecycle = "2.9.6"
 androidx-material = "1.13.0"
 androidx-test-junit = "1.2.1"
-compose-plugin = "1.9.3" # compose multiplatform plugin
+compose-plugin = "1.10.1" # compose multiplatform plugin
 junit = "4.13.2"
-kotlin = "2.3.0"
-ktorVersion = "3.3.3"
+kotlin = "2.3.10"
+ktorVersion = "3.4.1"
 coroutineVersion = "1.10.2"
-serializationVersion = "1.9.0"
+serializationVersion = "1.10.0"
 voyager = "1.1.0-beta03"
-kermit = "2.0.8"
+kermit = "2.1.0"
 koin = "4.1.1"
 sqlDelight = "2.2.1"
-coilVerion = "3.3.0"
-foundationLayoutAndroid = "1.10.0"
-firebaseDataconnect = "17.1.2"
+coilVerion = "3.4.0"
+foundationLayoutAndroid = "1.10.4"
+firebaseDataconnect = "17.1.4"
 androidXTestCore = "1.7.0"
 testRunner = "1.7.0"
-robolectric = "4.16"
+robolectric = "4.16.1"
 testExtJunit = "1.3.0"
 detekt = "1.23.8"
 composablesIcons = "2.2.1"
